+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Add PgConn.CleanupChan so connection pools can determine when async close is complete
+
 # 1.6.4 (July 29, 2020)
 
 * Fix deadlock on error after CommandComplete but before ReadyForQuery